requirements.txt
absl-py==2.0.0
appdirs==1.4.4
asttokens==2.4.0
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.0
flatbuffers==23.5.26
fonttools==4.43.1
frozendict==2.3.8
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.0
grpcio==1.59.0
h5py==3.10.0
html5lib==1.1
idna==3.4
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.4.0
jupyter_core==5.4.0
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
multitasking==0.0.11
nest-asyncio==1.5.8
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
patsy==0.5.3
peewee==3.16.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.21
stack-data==0.6.3
statsmodels==0.14.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
webencodings==0.5.1
Werkzeug==3.0.0
wrapt==1.14.1
yfinance==0.2.31
